+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 9.22.11 -
 ### Added
+- Added Cooking Decor
 - Added DF and SL Cooking recipes
 
 ## 9.22.10 - July 1, 2026